requirements/prod.txt
Django==4.2.9
django-cities-light==3.10.1
django_recaptcha==4.0.0
Pillow==10.3.0
python-dotenv==1.0.1
sorl-thumbnail==12.10.0
yookassa==3.1.0
